requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
amqp==5.3.1
attrs==24.2.0
Authlib==1.3.2
billiard==4.2.1
blinker==1.9.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.0
Flask==3.1.0
frozenlist==1.5.0
gunicorn==23.0.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.4.2
MarkupSafe==3.0.2
multidict==6.1.0
outcome==1.3.0.post0
packaging==24.2
prompt_toolkit==3.0.48
propcache==0.2.1
pycparser==2.22
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.0
requests==2.32.3
selenium==4.27.1
six==1.16.0
sniffio==1.3.1
some-package==0.1
sortedcontainers==2.4.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.18.3
